[FEATURE] Transparent Request passthrough as clone
When rendering Flux templates the Provider and therefore FluxService were unaware of the request being processed when called from the frontend. By wrapping the Request as a clone inside the ViewContext already used by FluxService we allow the request to be specified. And by making the Providers create the ViewContext with support for specifying the Request there's finally a link from the frontend controller's request straight through to rendering the Flux templates.