[BUGFIX] New items in Flux containers cannot be translated
When a new element was added to the original language version of an
already translated Flux container, the translate button was rendered in
the translated container. However, when attempting a translation, an
empty element list was returned, letting the process crash when
attempting the translation.

The cause of the error is that Flux does not properly pass on its
information to the core (i.e. the method that renders the "Translate"
button) and that the core does not support passing on additional
information about nested data structures in the translation process.

The fix is to pass on the correct colPos, so we can easily separate
translations inside Flux elements from the others, and pass the element
ids, which are already present in the core, but not passed through the JS
code.

This patch additionally requires a core patch!

Resolves: #20834